finalização operacao bancaria

+++ src/VilevePay.Application/ViewModels/v1/Parametrizacao/OperacaoBancariaViewModel.cs
@@ -2,7 +2,7 @@
 {
     public class OperacaoBancariaViewModel
     {
-        public int CodigoOperacao { get; set; }
+        public long CodigoOperacao { get; set; }
         public string Tipo { get; set; }
         public string Nome { get; set; }
     }
